Remove TicketCommand abstract class
