--- module/Application/src/Application/Command/Ticket/TicketCommand.php
@@ -1,47 +1,0 @@
-<?php
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license.
- */
-
-namespace Application\Command\Ticket;
-
-use Application\Entity\Ticket;
-
-abstract class TicketCommand
-{
-    /**
-     * @var Ticket
-     */
-    protected $ticketEntity;
-
-    public function __construct($subject, $description, $importance, $projectId, $openedBy)
-    {
-        $this->ticketEntity = new Ticket();
-        $this->ticketEntity->setSubject($subject);
-        $this->ticketEntity->setDescription($description);
-        $this->ticketEntity->setImportance($importance);
-        $this->ticketEntity->setProjectId($projectId);
-        $this->ticketEntity->setOpenedBy($openedBy);
-    }
-
-    /**
-     * @return Ticket
-     */
-    public function getEntity()
-    {
-        return $this->ticketEntity;
-    }
-}
